requirements.txt
azure-ai-documentintelligence == 1.0.0b4                 
azure-ai-vision-face == 1.0.0b1              
azure-cognitiveservices-nspkg == 3.0.1                  
azure-cognitiveservices-vision-computervision == 0.9.0                 
azure-cognitiveservices-vision-customvision == 3.1.0                   
azure-cognitiveservices-vision-face == 0.6.0                  
azure-cognitiveservices-vision-nspkg == 3.0.1                  
azure-common == 1.1.28                  
azure-core == 1.31.0                  
azure-identity == 1.18.0                
azure-nspkg == 3.0.2                   
azure-storage-blob == 12.23.1  
flask == 3.0.3   
gradio == 5.0.1                  
gradio-client == 1.4.0      
h11 == 0.14.0         
ipykernel == 6.29.5        
ipython == 8.27.0   
jinja2 == 3.1.4           
jupyter_client == 8.6.3   
jupyter_core == 5.7.2  
matplotlib == 3.9.2  
matplotlib-inline == 0.1.7  
numpy == 2.1.1 
pandas == 2.2.3  
pickleshare == 0.7.5            
pillow == 10.4.0    
pip == 24.2  
psutil == 5.9.0 
pydantic == 2.9.2  
pydantic-core == 2.23.4 
python == 3.12.3  
python-dateutil == 2.9.0      
python-dotenv == 1.0.1  
pyyaml == 6.0.2   
requests == 2.32.3 
setuptools == 72.1.0  
tabulate2 == 1.9.1    
video-indexer == 0.1.8 
websockets == 12.0  
yaml == 0.2.5              